Implement openRoute on gateway, create support for intermediaries
This creates an implementation of the openRoute function on gateways,
which is useful for when clients can not directly arbitrate a
connection. It is also required in order to ensure full functionality.
This also enhances the client's support for handling OPENROUTE messages
by allowing them to utilize intermediaries to the final destination.
This has not been tested, and will almost certainly undergo some
changes.