Fixed Address Parsing Bug
Fixed bug with byte order when stringifying. Also fixed another bug related to array indexes. Somehow the arrays lined up so that the error was not obvious nor randomized in effect (i was before e and e+1 was always null?)

Address handling was not affected by this bug only string outs.