Oh Lordy
Lots of mostly runtime patches.
Patched GERTe library to correctly execute the route cleanup rather than try to delete all the keys.
Patched the keys to correctly export the key database management functions.
Patched connections to initialize state to FAILURE. This is used to signal the library that it's a new connection.
Patched the gateway address grabber to correctly set itself instead of causing weird "this pointer is totes a socket" errors.