General Fixes
Added extra error logging on Windows (as it supports it.) This error logging does cause a minor memory leak sadly but that is being actively debugged.
Fixed polls not being updated when sockets are closed in certain scenarios.
Removed a previous cleanup due to a fix that was believed to be unneeded actually being needed.