Fixed Linux Crashing
Fixed Linux crashing by using SOCKETs to reference the current object instead.

Cause of the crash analysis:
In order to post-construct the registered gateway, the unregistered gateway was moved within the newly allocated Gateway object. This move occurred before any form of network poll updating. This caused the network poll to incorrectly read an uninitialized portion of memory to find and remove or add a socket. To fix this, the socket is independently stored with the thread dependent on it.