Improved Code Security
Several compiler settings have been changed so more explicitly define how the compiler interprets the code on Windows. This results in a much stricter compiler (such as using the SDL flag, which was activated) which turned certain warnings on, and other warnings into errors. This commit fixes those warnings/errors.