GERTi v1.2 Build 2
Changed the connection infrastructure to refer to Connection Indexes generated when a connection is formed instead of using the old separate paths and connections tables. This allowed me to simplify 2 3-deep tables into 1 2-deep table.

Changed GERTe behavior slightly. The MNC will now check to see if the GERTe address of the destination is different from the GERTe address of the MNC. This should reduce unnecessary Internet traffic.

Added "getVersion" which will return a string identifying the GERTi version currently in use.

Prepared MNC for supporting more than 3 tiers.