v1.4.1 Fix (#79)
* GERTi v1.2.1 with feeling

Finish properly sending arguments in GERTi.send and update the version number

* GERTi v1.2.2

I hate Lua dynamic typing. Fixes a problem with floats that should be integers.

* Add an RC version of GERTiMNC

Adds a RC version of GERTiMNC for an alternative method of autoloading it on boot and having it run in the background

* Rename file

Apparently, GitHub only loads Readme.md

* GERTi v1.3 Build 1

Begin cleanup of functions and start shrinking size. At this point, still compatible with v1.2, but this will change.

Completely untested and likely contains bugs.

* GERTi v1.3 - Build 2

Continued cleanup of the route opening logic and the startup logic. This is now protocol incompatible with v1.2.

Fixed at least one bug in Build 1, but still untested and likely broken.

* GERTiClient v1.3 - Build 3

Finish all the theoretical changes to GERTiClient to simplify route opening from 3 functions down to 2. Socket opening changed to reflect this.

Only remaining work is bugtesting and GERTiMNC feature addition

* GERTiMNC v1.3 - Build 3

-- Change: GERTiMNC now solely resides as an RC service in /etc/rc.d

--Change: Configuration files have moved to /etc

-- Feature Addition: triggering the loadAddress command will have the MNC reload the list of addresses. This way, you can change the modem address tied to a certain GERTi address without restarting the entire network. The GERTiClient for that node must be restarted for it to take effect.

* GERTi v1.3 - Build 4

Fix some errors that result in the Client failing to boot or obtain an address. More errors likely await in connection handling.

* GERTiClient v1.3 - Build 5

Now comes with the ability to open connections at least to neighbors. With the exception of GERTi.send(), data transmission is still bugged

* GERTi v1.3 - Build 6

Data transmission now works between neighbors at least. Final testing will involve linked cards and multiple hops.

* GERTi v1.3 - Build 7

Fixes a bug that prevents users from closing connections

* GERTi v1.3 Release

After testing and bugfixing, GERTi is now stable. Protocol incompatible with v1.2, but API compatible

* Update Reserved GERTe numbers.txt

* Update Reserved GERTe numbers.txt

* Cleanup

* GERTeAPI v1.3 - Build 1

Update the GERTeAPI to fit the GERTiMNC RC daemon. Pending testing.

* GERTi v1.3.1

Update GERTiMNC to resolve performance drop with incoming GERTe packets

* GERTi v1.4 Build 1

Lifts the linked and modem card combination limits. Completely untested and likely has bugs

* GERTi v1.4 Build 2

More bugfixes. Now it will
1) Actually load
2) Support multiple cards on the MNC properly
3) Actually open connections

* GERTi v1.4 Build 3

All known currently existing bugs have been patched. This should be suitable for temporary deployment. Subsequent builds will likely include more bugs.

* GERTi v1.4 RC 1

This addition includes all planned features and has been bugtested. It is suitable for deployment across networks.

Release candidate instead of final only because a slight change to Data handling may occur where computers may be able to silently re-add connections to their connection tables, thus allowing for more graceful handling of hard power restarts.

* GERTi v1.4 - Release

Fixed a bug with GERTiMNC address assignment
Added a feature allowing the receiving end of a connection to automatically regenerate the connection if it reboots improperly without executing the shutdown or reboot command.

* GERTi v1.4.1 - Build 1

Hopefully a fix to improper addressing with temporary debug statements

* GERTi v1.4.1 Build 2

Fixes a problem with connections and hairpin turns. Slightly improves performance when a route would originally have to pass through a computer, go to the MNC, and then pass back through the computer.

* GERTi v1.4.1 - Build 3

Fixes the stupid addressing error once and for all. If I have to fix this again, it's going to be an issue

* Update GERTiClient.lua (#76)

This is my proposed bugfix to GERTiClient, which solves the issues I was encountering on the test world. Due to the nature of the changes, they should have no unintended consequences on how GERTiClient functions on clients that they already worked on before, but will allow it to function on the network I have tested, as well as ensuring variable type in the future for the Connection ID.

* Fix Improper Handling of cPend Table

When a destination of a connection receives the connection open request, it still creates an entry in the cPend table, which should not happen. This fixes that error.

* GERT v1.4.1 Build 5

Vastly revises how the MNC stores and handles the network nodes table. The table will now properly update neighbor connections and handle nodes disconnecting from the network. This stops the gradual divergence of what the MNC sees of the network from how the network is actually laid out.

* GERTi v1.4.1 Build 6

Further changes to the structure of the neighbors table and final preparation of the routing code changes

* GERTi v1.4.1 Release

Completes rework of MNC routing recode for greater simplicity and sensibility

* GERTi v1.4.1 True Release

Fixes a bug in removing nodes from the network

* v1.4.1 Repair

Fix something broken by the GitHub merge

Co-authored-by: leothehero <leothehero@users.noreply.github.com>