The Cleanliness Update
Standardized errors codes across both GUSs and GUSc.
Added automatic downloading of updates on GUSc.

Automatic Updates are now a toggleable config option for both GUSs and GUSc, the config option is called "DailyCheck", is true by default, and can be modified and viewed with the ChangeConfigSetting() and GetSetting() methods which are now available on both.

Additionally, GUSs has a new config setting: "StartImmediately", which determines whether or not the program starts the event listeners on first load. It is true by default.