GERTUpdateServer DNS update
A new config option has been added to GUSs: "SETDNS", which is true by default. If done so, it registers itself under the "GUS" hostname. This allows it to be moved off of the MNC, granted the DNS service is operational. If the DNS service is not operational, GUSc will fallback to 0.0 as the address, and any GUSs not on the MNC will not function.