Gus 1.2 (#102)
* The Cleanliness Update

Standardized errors codes across both GUSs and GUSc.
Added automatic downloading of updates on GUSc.

Automatic Updates are now a toggleable config option for both GUSs and GUSc, the config option is called "DailyCheck", is true by default, and can be modified and viewed with the ChangeConfigSetting() and GetSetting() methods which are now available on both.

Additionally, GUSs has a new config setting: "StartImmediately", which determines whether or not the program starts the event listeners on first load. It is true by default.

* GERTUpdateServer DNS update

A new config option has been added to GUSs: "SETDNS", which is true by default. If done so, it registers itself under the "GUS" hostname. This allows it to be moved off of the MNC, granted the DNS service is operational. If the DNS service is not operational, GUSc will fallback to 0.0 as the address, and any GUSs not on the MNC will not function.

* Update RecommendedConfig-GUSs.cfg