GERTi v1.5.1 Build 1
GERTi will now notify users about the usage of Lua 5.2, which is discouraged on clients and forbidden on the MNC

GERTi.flushDNS has been added to allow programs to clear GERTiClient's DNS cache. On the MNC, this will erase every single DNS binding on the network.

The MNC will remember previously registered modules, and will attempt to reload them on startup.